+++ src/components/RegisterForm/RegisterForm.module.scss
@@ -7,6 +7,8 @@
 
 
 .submit {
+    width: 100%;
+    
     & .spinner {
         stroke: white;
     }
